Code changed to enable both the input1 and input2 control lines of head unit. As per the change, all double clicks will be sent through input2 line and all single clicks would be sent through input1 line.
While sending a command to HU, the command value and the input line will be
selected according to the click type.

This code now expect the dual digipot, in which digipot 1 will be connected
to input1 and digipot2 will be connected to input2. These lines can be enabled
indipendently. However the circuit will be made with a jumper setting to
choose whether input2 relay should depend  on input1 relay or not.Here the
dependency means, input2 will be activated only when input1 is activated.

As per the  current change, the code expects that the jumper is set to make
input2 independant of input1. So that input2 line can be enabled without
enabling input1 relay.

So, when a single click happens, the command will be fetched from the
indexmap[click_type][key_value] location, and the digipot and relay for
input line will be selected uing the click_type.

In case of macro, now the macro step data type is chnaged from uint8_t to
uint16_t so that the MSB of step stores the click type(0 or 1) and the LSB
of the step stores the command to send. With this data, each macro step can
be sent through the corresponding input lines according to the click type.

This change is compiled successfully but not loaded and tested. Once it is
tested, there may be more commits to make this functionality bug free.